Gamma perspective added
