+++ plugins/core/hu.bme.mit.gamma.gui/META-INF/MANIFEST.MF
@@ -3,11 +3,13 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Gamma Graphical User Interface
 Bundle-SymbolicName: hu.bme.mit.gamma.gui;singleton:=true
 Bundle-Version: 2.8.0.qualifier
-Export-Package: hu.bme.mit.gamma.gui
+Export-Package: hu.bme.mit.gamma.gui.perspectives,
+ hu.bme.mit.gamma.gui.wizards
 Require-Bundle: org.eclipse.jface,
  org.eclipse.ui.workbench,
  org.eclipse.equinox.registry,
  org.eclipse.core.resources,
- org.eclipse.ui.ide
+ org.eclipse.ui.ide,
+ org.eclipse.ui.console
 Automatic-Module-Name: hu.bme.mit.gamma.gui
 Bundle-RequiredExecutionEnvironment: JavaSE-17
