+++ plugins/core/hu.bme.mit.gamma.gui/build.properties
@@ -1,4 +1,8 @@
 source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
-               .
+               .,\
+               src/,\
+               icons/,\
+               plugin.xml,\
+               bin/
