+++ plugins/core/hu.bme.mit.gamma.gui/src/hu/bme/mit/gamma/gui/wizards/NewGCLInterfacesWizard.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-1.0
  ********************************************************************************/
-package hu.bme.mit.gamma.gui;
+package hu.bme.mit.gamma.gui.wizards;
 
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.wizard.Wizard;
