+++ plugins/core/hu.bme.mit.gamma.gui/src/hu/bme/mit/gamma/gui/wizards/NewGCLStatechartWizardWizardPage.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-1.0
  ********************************************************************************/
-package hu.bme.mit.gamma.gui;
+package hu.bme.mit.gamma.gui.wizards;
 
 import org.eclipse.core.resources.IContainer;
 import org.eclipse.core.resources.IResource;
