+++ plugins/rcp/hu.bme.mit.gamma.rcp/META-INF/MANIFEST.MF
@@ -98,7 +98,8 @@ Require-Bundle: javax.inject;bundle-version="0.0.0",
  hu.bme.mit.gamma.yakindu.interface.transformation.commandhandler;bundle-version="2.8.0",
  hu.bme.mit.gamma.yakindu.transformation;bundle-version="2.8.0",
  hu.bme.mit.gamma.yakindu.transformation.traceability;bundle-version="2.8.0",
- org.eclipse.ui.intro
+ org.eclipse.ui.intro,
+ hu.bme.mit.gamma.gui;bundle-version="2.8.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Import-Package: javax.annotation;version="0.0.0"
 Automatic-Module-Name: hu.bme.mit.gamma.rcp
