+++ plugins/rcp/hu.bme.mit.gamma.rcp/plugin_customization.ini
@@ -1,1 +1,4 @@
 org.eclipse.ui/SHOW_PROGRESS_ON_STARTUP = true
+org.eclipse.ui/defaultPerspectiveId=hu.bme.mit.gamma.gui.perspective
+org.eclipse.ui.workbench/org.eclipse.ui.commands=<?xml version\="1.0" encoding\="UTF-8"?>\r\n<org.eclipse.ui.commands>\r\n<keyBinding contextId\="org.eclipse.ui.contexts.window" keyConfigurationId\="org.eclipse.ui.defaultAcceleratorConfiguration" keySequence\="ALT+CTRL+G"/>\r\n</org.eclipse.ui.commands>
+
